test_compiler.sh is not user friendly! Usage:
    ./test_compiler.sh -b
    ^ Run all bad tests (These should fail). They are found in test_programs/bad_programs/*

    ./test_compiler.sh -g
    ^ Run all good tests (These should fail). They are found in test_programs/good_programs/*

    ./test_compiler.sh -a
    ^ Run all tests

    ./test_compiler.sh
    ^ Usage info

A note on how to write tests:

Good Tests
==========
These should all be valid ibtl code. They are stored in 'test_programs/good_programs/'.

All good tests should have two files with the following naming convetion:

    X.ibtl
    X.expect

The test file with the 'ibtl' extension has ibtl code and the file with the extention 'expect' should have what the code compiled by the compiler should look like when you run it with the `gforth` binary.

Bad Tests
=========
These should be ibtl that is symantically incorrect. They are stored in 'test_programs/bad_programs/'.

All bad tests are in the following format.

    X.ibtl

Note that there is no file with the 'expect' extention.